refactoring : 2D 렌더링 기능 제거

+++ MyEngine_SOURCE/MyVertexBuffer.cpp
@@ -13,7 +13,7 @@ namespace Source::Graphics
 
 	bool VertexBuffer::Create(const std::vector<Vertex>& vertexes)
 	{
-		Description.ByteWidth = sizeof(Vertex) * vertexes.size();
+		Description.ByteWidth = sizeof(Vertex) * (UINT)vertexes.size();
 		Description.BindFlags = D3D11_BIND_FLAG::D3D11_BIND_VERTEX_BUFFER;
 		Description.Usage = D3D11_USAGE::D3D11_USAGE_DYNAMIC;
 		Description.CPUAccessFlags = D3D11_CPU_ACCESS_FLAG::D3D11_CPU_ACCESS_WRITE;
